Convert form into modal
