+++ app/client/src/pages/Editor/CurlImport/CurlImportForm.tsx
@@ -3,19 +3,15 @@ import type { InjectedFormProps } from "redux-form";
 import { reduxForm, Form, Field } from "redux-form";
 import styled from "styled-components";
 import { CURL_IMPORT_FORM } from "@appsmith/constants/forms";
-import type { curlImportFormValues } from "./helpers";
-import { curlImportSubmitHandler } from "./helpers";
 import CurlLogo from "assets/images/Curl-logo.svg";
-import { Button } from "design-system";
-import FormRow from "components/editorComponents/FormRow";
-import Debugger, {
-  ResizerContentContainer,
-  ResizerMainContainer,
-} from "../DataSourceEditor/Debugger";
-
-const MainConfiguration = styled.div`
-  padding: var(--ads-v2-spaces-4) var(--ads-v2-spaces-7);
-`;
+import {
+  Button,
+  Modal,
+  ModalContent,
+  ModalFooter,
+  ModalHeader,
+} from "design-system";
+import { type curlImportFormValues, curlImportSubmitHandler } from "./helpers";
 
 const ActionButtons = styled.div`
   justify-self: flex-end;
@@ -44,7 +40,6 @@ const CurlImportText = styled.p`
   font-size: 17px;
   font-weight: 500;
   color: var(--ads-v2-color-fg);
-  font-size: 17px;
   line-height: 22px;
   letter-spacing: -0.204px;
 `;
@@ -80,16 +75,6 @@ const CurlImportFormContainer = styled.div`
     font-size: ${(props) => props.theme.fontSizes[3]}px;
   }
 `;
-const MainContainer = styled.div`
-  display: flex;
-  position: relative;
-  height: 100%;
-  flex-direction: column;
-  /* padding: var(--ads-v2-spaces-7); */
-  .curl-form-resizer-content {
-    padding: 0px var(--ads-v2-spaces-7);
-  }
-`;
 
 interface OwnProps {
   isImportingCurl: boolean;
@@ -99,45 +84,25 @@ interface OwnProps {
     dispatch: any,
   ) => void;
   initialValues: Record<string, unknown>;
-  closeEditorLink?: React.ReactNode;
 }
 
 type Props = OwnProps & InjectedFormProps<curlImportFormValues, OwnProps>;
 
 class CurlImportForm extends React.Component<Props> {
   render() {
-    const { closeEditorLink, handleSubmit, isImportingCurl, showDebugger } =
-      this.props;
+    const { handleSubmit, isImportingCurl } = this.props;
 
     return (
-      <MainContainer>
-        {closeEditorLink}
-        <MainConfiguration>
-          <FormRow className="form-row-header">
-            <div
-              style={{
-                display: "flex",
-              }}
-            >
-              <CurlIconWrapper>
-                <img alt="CURL" src={CurlLogo} />
-              </CurlIconWrapper>
-              <CurlImportText className="text">Import from CURL</CurlImportText>
-            </div>
-            <ActionButtons className="t--formActionButtons">
-              <Button
-                className="t--importBtn"
-                isLoading={isImportingCurl}
-                onClick={handleSubmit(curlImportSubmitHandler)}
-                size="md"
-              >
-                Import
-              </Button>
-            </ActionButtons>
-          </FormRow>
-        </MainConfiguration>
-        <ResizerMainContainer>
-          <ResizerContentContainer className="curl-form-resizer-content">
+      <Modal open>
+        <ModalContent>
+          <ModalHeader>
+            <CurlIconWrapper>
+              <img alt="CURL" src={CurlLogo} />
+            </CurlIconWrapper>
+            <CurlImportText className="text">Import from CURL</CurlImportText>
+          </ModalHeader>
+
+          <div className="curl-form-resizer-content">
             <StyledForm onSubmit={handleSubmit(curlImportSubmitHandler)}>
               <label className="inputLabel">Paste CURL Code Here</label>
               <CurlHintText>
@@ -157,10 +122,21 @@ class CurlImportForm extends React.Component<Props> {
                 <Field component="input" name="name" type="hidden" />
               </CurlImportFormContainer>
             </StyledForm>
-          </ResizerContentContainer>
-          {showDebugger && <Debugger />}
-        </ResizerMainContainer>
-      </MainContainer>
+          </div>
+          <ModalFooter>
+            <ActionButtons className="t--formActionButtons">
+              <Button
+                className="t--importBtn"
+                isLoading={isImportingCurl}
+                onClick={handleSubmit(curlImportSubmitHandler)}
+                size="md"
+              >
+                Import
+              </Button>
+            </ActionButtons>
+          </ModalFooter>
+        </ModalContent>
+      </Modal>
     );
   }
 }
