+++ app/client/src/pages/Editor/CurlImport/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CurlImportEditor";
